Consolidate Firebase service into Firebase client
The concept of a service isn’t strongly motivated in Popcode—the idea
primarily applies to dependency injection frameworks, which we don’t
use. Instead, just have the client be responsible for maintaining
instances of the underlying SDK as needed. `gapi` should probably go
this way as well.
